third_party/WebKit/LayoutTests/media/video-dom-preload.html

<!DOCTYPE html>
<title>Test media "autoplay" attribute set via DOM.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
test(function() {
    var video = document.querySelector("video");

    // Test default attribute value.
    assert_equals(video.preload, "auto");
    assert_equals(video.getAttribute("preload"), null);

    // Remove attribute, should revert to default.
    video.removeAttribute("preload");
    assert_equals(video.preload, "auto");

    checkPreloadValue("none", "none");
    checkPreloadValue("auto", "auto");

    // set to bogus value, should revert to default value.
    video.removeAttribute("preload");
    // Set via IDL attribute
    video.preload = "bogus";
    assert_equals(video.getAttribute("preload"), "bogus");
    assert_equals(video.preload, "auto");
    // - and via content attribute.
    video.setAttribute("preload", "bogus");
    assert_equals(video.preload, "auto");
    assert_equals(video.getAttribute("preload"), "bogus");

    checkPreloadValue("metadata", "metadata");

    function checkPreloadValue(value, expected) {
        // Set via IDL attribute
        video.removeAttribute("preload");
        video.preload = value;
        assert_equals(video.getAttribute("preload"), expected);
        assert_equals(video.preload, expected);
        // - and via content attribute.
        video.removeAttribute("preload");
        video.setAttribute("preload", value);
        assert_equals(video.preload, expected);
        assert_equals(video.getAttribute("preload"), expected);
    }
});
</script>
